+++ src/main/java/me/banbeucmas/oregen3/listener/BlockListener.java
@@ -127,7 +127,7 @@ public class BlockListener implements Listener {
         if (finalDelay > 0) {
             OfflinePlayer owner = plugin.getUtils().getOwner(block.getLocation());
             if (owner != null) {
-                int level = plugin.getUpgradeManager().getUpgradeLevel(owner.getUniqueId(), "linhmach");
+                double level = plugin.getUpgradeManager().getUpgradeLevel(owner.getUniqueId(), "linhmach");
                 if (level > 0) {
                     double percentPerLevel = config.getDouble("upgrade-gui.upgrades.linhmach.percent_per_level", 5.0);
                     double speedFactor = 1.0 + (percentPerLevel * level / 100.0);
